+++ CHANGELOG.md
@@ -1,2 +1,7 @@
-﻿# 0.1.0
+﻿# 0.2.0
+Changes:
+- Deprecate Canvas field in Controller. Use List<Canvas> Canvases field instead to allow a scene to has multiple Canvases.
+- If scene's canvas's sorting order is set to a value different from 0, its sorting order will not be changed to the scene's sorting order.
+
+# 0.1.0
 - First version.